+++ packages/compass-components/src/components/actions/constants.tsx
@@ -1,0 +1,8 @@
+export const ItemActionButtonSize = {
+  XSmall: 'xsmall',
+  Small: 'small',
+  Default: 'default',
+} as const;
+
+export type ItemActionButtonSize =
+  typeof ItemActionButtonSize[keyof typeof ItemActionButtonSize];
