+++ packages/compass-components/src/components/actions/item-action-controls.tsx
@@ -1,96 +1,17 @@
-import React, {
-  useRef,
-  forwardRef,
-  useCallback,
-  useState,
-  useMemo,
-} from 'react';
-import {
-  Button,
-  Icon,
-  IconButton,
-  Menu,
-  MenuItem,
-  MenuSeparator,
-  Tooltip,
-} from '../leafygreen';
-import type { ButtonProps } from '@leafygreen-ui/button';
-import type { glyphs } from '@leafygreen-ui/icon';
+import React, { useMemo } from 'react';
 import { spacing } from '@leafygreen-ui/tokens';
 import { css, cx } from '@leafygreen-ui/emotion';
-import { WorkspaceContainer } from '../workspace-container';
 
-type ItemBase<Action extends string> = {
-  action: Action;
-  label: string;
-  icon?: React.ReactChild;
-  variant?: 'default' | 'destructive';
-  isDisabled?: boolean;
-  disabledDescription?: string;
-  tooltip?: string;
-  className?: string;
-  /**
-   * How to show the item when not collapsed into the menu.
-   * @default ItemActionButton
-   */
-  expandedAs?: React.ComponentType<ItemComponentProps<Action>>;
-};
-
-export type ItemComponentProps<Action extends string> = Omit<
-  ItemBase<Action>,
-  'expandedAs'
-> & {
-  // TODO: Rename to "size"
-  iconSize?: ItemActionButtonSize;
-  iconClassName?: string;
-  iconStyle?: React.CSSProperties;
-  'data-testid'?: string;
-  onClick(evt: React.MouseEvent<unknown>): void;
-};
-
-export type ItemAction<Action extends string> = {
-  icon: keyof typeof glyphs | React.ReactElement;
-} & ItemBase<Action>;
-
-export type ItemSeparator = { separator: true };
-
-export type GroupedItemAction<Action extends string> = ItemAction<Action> & {
-  tooltipProps?: Parameters<typeof Tooltip>;
-};
-
-export type MenuAction<Action extends string> =
-  | ItemBase<Action>
-  | ItemSeparator;
-
-function isSeparatorMenuAction(value: unknown): value is ItemSeparator {
-  return (
-    typeof value === 'object' &&
-    value !== null &&
-    'separator' in value &&
-    value.separator === true
-  );
-}
-
-const ItemActionButtonSize = {
-  XSmall: 'xsmall',
-  Small: 'small',
-  Default: 'default',
-} as const;
-
-type ItemActionButtonSize =
-  typeof ItemActionButtonSize[keyof typeof ItemActionButtonSize];
+import { ItemActionMenu } from './item-action-menu';
+import { ItemActionButtonSize } from './constants';
+import type { ItemAction, ItemSeparator } from './types';
+import { ItemActionGroup } from './item-action-group';
 
 const actionControlsStyle = css({
   flex: 'none',
   marginLeft: 'auto',
   alignItems: 'center',
   display: 'flex',
-});
-
-const actionGroupButtonStyle = css({
-  '&:not(:first-child)': {
-    marginLeft: spacing[100],
-  },
 });
 
 // Action buttons are rendered 4px apart from each other. With this we keep the
@@ -100,344 +21,7 @@ const actionMenuWithActionControlsStyles = css({
   marginLeft: spacing[100],
 });
 
-const iconContainerStyle = css({
-  display: 'block',
-  flex: 'none',
-  fontSize: 0,
-  lineHeight: 0,
-});
-
-// Using important here because leafygreen / emotion applies styles in the order
-// that doesn't allow our styles override theirs
-const buttonSizeStyle: Record<ItemActionButtonSize, string | undefined> = {
-  default: undefined,
-  small: css({
-    flex: 'none',
-    width: `${spacing[4]}px !important`,
-    height: `${spacing[4]}px !important`,
-  }),
-  xsmall: css({
-    flex: 'none',
-    // aligns with other xsmall components
-    width: `${20}px !important`,
-    height: `${20}px !important`,
-  }),
-};
-
-function actionTestId<Action extends string>(
-  dataTestId: string | undefined,
-  action: Action
-) {
-  return dataTestId ? `${dataTestId}-${action}-action` : undefined;
-}
-
-// As we are using this component to render icon in MenuItem,
-// and it does cloneElement on glyph, here we are accepting all the
-// props that are passed during clone process.
-type IconProps = React.ComponentProps<typeof Icon>;
-const ActionGlyph = ({
-  glyph,
-  size,
-  ...props
-}: Omit<IconProps, 'size' | 'glyph'> & {
-  glyph?: React.ReactChild;
-  size?: ItemActionButtonSize;
-}) => {
-  if (typeof glyph === 'string') {
-    return <Icon size={size} glyph={glyph} {...props} />;
-  }
-
-  if (React.isValidElement(glyph)) {
-    return glyph;
-  }
-
-  return null;
-};
-
-const SmallIconButton = forwardRef<
-  HTMLButtonElement,
-  {
-    glyph: React.ReactChild;
-    label: string;
-    title?: string;
-    size: ItemActionButtonSize;
-    onClick(evt: React.MouseEvent<HTMLButtonElement>): void;
-  } & Omit<React.HTMLProps<HTMLButtonElement>, 'size'>
->(function SmallIconButton(
-  { glyph, size, label, onClick, children, title, className, ...rest },
-  ref
-) {
-  return (
-    <IconButton
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-expect-error leafygreen confuses TS a lot here
-      ref={ref}
-      className={cx(buttonSizeStyle[size], className)}
-      aria-label={label}
-      title={title}
-      onClick={onClick}
-      {...rest}
-    >
-      <span role="presentation" className={iconContainerStyle}>
-        <ActionGlyph glyph={glyph} size={size} />
-      </span>
-      {/* Only here to make leafygreen menus work */}
-      {children}
-    </IconButton>
-  );
-});
-
-export function ItemActionButton<Action extends string>({
-  action,
-  icon = <></>,
-  label,
-  tooltip,
-  iconSize = ItemActionButtonSize.Default,
-  onClick,
-  iconClassName,
-  className,
-  iconStyle,
-  isDisabled,
-  'data-testid': dataTestId,
-}: ItemComponentProps<Action>) {
-  return (
-    <SmallIconButton
-      key={action}
-      glyph={icon}
-      label={label}
-      title={!tooltip ? label : undefined}
-      size={iconSize}
-      data-action={action}
-      data-testid={dataTestId}
-      onClick={onClick}
-      className={cx(actionGroupButtonStyle, iconClassName, className)}
-      style={iconStyle}
-      disabled={isDisabled}
-    />
-  );
-}
-
-export function ItemActionMenu<Action extends string>({
-  isVisible = true,
-  actions,
-  onAction,
-  className,
-  menuClassName,
-  usePortal,
-  iconClassName,
-  iconStyle,
-  iconSize = ItemActionButtonSize.Default,
-  'data-testid': dataTestId,
-}: {
-  actions: MenuAction<Action>[];
-  onAction(actionName: Action): void;
-  className?: string;
-  menuClassName?: string;
-  usePortal?: boolean;
-  iconClassName?: string;
-  iconStyle?: React.CSSProperties;
-  iconSize?: ItemActionButtonSize;
-  isVisible?: boolean;
-  'data-testid'?: string;
-}) {
-  // this ref is used by the Menu component to calculate the height and position
-  // of the menu, and by us to give back the focus to the trigger when the menu
-  // is closed (https://jira.mongodb.org/browse/PD-1674).
-  const menuTriggerRef = useRef<HTMLButtonElement | null>(null);
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-  const onClick = useCallback(
-    (evt) => {
-      evt.stopPropagation();
-      if (evt.currentTarget.dataset.menuitem) {
-        setIsMenuOpen(false);
-        // Workaround for https://jira.mongodb.org/browse/PD-1674
-        menuTriggerRef.current?.focus();
-      }
-      onAction(evt.currentTarget.dataset.action);
-    },
-    [onAction]
-  );
-
-  const shouldRender = isMenuOpen || (isVisible && actions.length > 0);
-
-  if (!shouldRender) {
-    return null;
-  }
-
-  return (
-    <div className={cx(actionControlsStyle, className)}>
-      <Menu
-        className={menuClassName}
-        open={isMenuOpen}
-        setOpen={setIsMenuOpen}
-        refEl={menuTriggerRef}
-        usePortal={usePortal}
-        data-testid={dataTestId}
-        trigger={({
-          onClick,
-          children,
-        }: {
-          onClick: React.MouseEventHandler<HTMLButtonElement>;
-          children: React.ReactNode;
-        }) => {
-          return (
-            <SmallIconButton
-              ref={menuTriggerRef}
-              size={iconSize}
-              glyph="Ellipsis"
-              label="Show actions"
-              title="Show actions"
-              data-testid={
-                dataTestId ? `${dataTestId}-show-actions` : undefined
-              }
-              onClick={(evt) => {
-                evt.stopPropagation();
-                onClick && onClick(evt);
-              }}
-              className={cx(actionGroupButtonStyle, iconClassName)}
-              style={iconStyle}
-            >
-              {children}
-            </SmallIconButton>
-          );
-        }}
-      >
-        {actions.map((menuAction, idx) => {
-          if (isSeparatorMenuAction(menuAction)) {
-            return <MenuSeparator key={`separator-${idx}`} />;
-          }
-
-          const {
-            action,
-            label,
-            icon,
-            variant,
-            isDisabled,
-            disabledDescription,
-          } = menuAction;
-
-          return (
-            <MenuItem
-              key={action}
-              data-testid={actionTestId<Action>(dataTestId, action)}
-              data-action={action}
-              data-menuitem={true}
-              glyph={<ActionGlyph glyph={icon} size={iconSize} />}
-              onClick={onClick}
-              variant={variant || 'default'}
-              disabled={isDisabled}
-              description={isDisabled ? disabledDescription : ''}
-            >
-              {label}
-            </MenuItem>
-          );
-        })}
-      </Menu>
-    </div>
-  );
-}
-
-export function ItemActionGroup<Action extends string>({
-  actions,
-  onAction,
-  className,
-  iconClassName,
-  iconStyle,
-  iconSize = ItemActionButtonSize.Default,
-  isVisible = true,
-  'data-testid': dataTestId,
-}: {
-  actions: (GroupedItemAction<Action> | ItemSeparator)[];
-  onAction(actionName: Action): void;
-  className?: string;
-  iconClassName?: string;
-  iconStyle?: React.CSSProperties;
-  iconSize?: ItemActionButtonSize;
-  isVisible?: boolean;
-  'data-testid'?: string;
-}) {
-  const onClick = useCallback(
-    (evt) => {
-      evt.stopPropagation();
-      onAction(evt.currentTarget.dataset.action);
-    },
-    [onAction]
-  );
-
-  const shouldRender = isVisible && actions.length > 0;
-
-  if (!shouldRender) {
-    return null;
-  }
-
-  return (
-    <div
-      className={cx(actionControlsStyle, className)}
-      data-testid={dataTestId}
-    >
-      {actions.map((menuItem, idx) => {
-        if (isSeparatorMenuAction(menuItem)) {
-          return <MenuSeparator key={`separator-${idx}`} />;
-        }
-
-        const {
-          expandedAs: ItemComponent = ItemActionButton,
-          tooltip,
-          tooltipProps,
-          ...itemProps
-        } = menuItem;
-
-        const item = (
-          <ItemComponent
-            {...itemProps}
-            iconSize={iconSize}
-            iconStyle={iconStyle}
-            iconClassName={iconClassName}
-            onClick={onClick}
-            data-testid={actionTestId<Action>(dataTestId, itemProps.action)}
-          />
-        );
-
-        if (tooltip) {
-          return (
-            <Tooltip
-              key={itemProps.action}
-              {...tooltipProps}
-              trigger={
-                <div
-                  className={actionGroupButtonStyle}
-                  style={{ display: 'inherit' }}
-                >
-                  {item}
-                </div>
-              }
-            >
-              {tooltip}
-            </Tooltip>
-          );
-        }
-
-        return item;
-      })}
-    </div>
-  );
-}
-
-export function ItemActionControls<Action extends string>({
-  isVisible = true,
-  actions,
-  onAction,
-  className,
-  menuClassName,
-  iconClassName,
-  iconStyle,
-  iconSize = ItemActionButtonSize.Default,
-  usePortal,
-  collapseAfter = 0,
-  collapseToMenuThreshold = 2,
-  'data-testid': dataTestId,
-}: {
+export type ItemActionControlsProps<Action extends string> = {
   isVisible?: boolean;
   actions: (ItemAction<Action> | ItemSeparator)[];
   onAction(actionName: Action): void;
@@ -452,7 +36,22 @@ export function ItemActionControls<Action extends string>({
   collapseToMenuThreshold?: number;
   usePortal?: boolean;
   'data-testid'?: string;
-}) {
+};
+
+export function ItemActionControls<Action extends string>({
+  isVisible = true,
+  actions,
+  onAction,
+  className,
+  menuClassName,
+  iconClassName,
+  iconStyle,
+  iconSize = ItemActionButtonSize.Default,
+  usePortal,
+  collapseAfter = 0,
+  collapseToMenuThreshold = 2,
+  'data-testid': dataTestId,
+}: ItemActionControlsProps<Action>) {
   const sharedProps = useMemo(
     () => ({
       isVisible,
@@ -520,118 +119,4 @@ export function ItemActionControls<Action extends string>({
   }
 
   return <ItemActionGroup actions={actions} {...sharedProps}></ItemActionGroup>;
-}
-
-const hiddenOnNarrowStyles = css({
-  [`@container ${WorkspaceContainer.toolbarContainerQueryName} (width < 900px)`]:
-    {
-      display: 'none',
-    },
-});
-
-export function DropdownMenuButton<Action extends string>({
-  isVisible = true,
-  actions,
-  onAction,
-  usePortal,
-  activeAction,
-  buttonText,
-  buttonProps,
-  iconSize = ItemActionButtonSize.Default,
-  'data-testid': dataTestId,
-  hideOnNarrow = true,
-}: {
-  actions: MenuAction<Action>[];
-  onAction(actionName: Action): void;
-  usePortal?: boolean;
-  iconSize?: ItemActionButtonSize;
-  isVisible?: boolean;
-  activeAction?: Action;
-  'data-testid'?: string;
-  buttonText: string;
-  buttonProps: ButtonProps;
-  hideOnNarrow?: boolean;
-}) {
-  // this ref is used by the Menu component to calculate the height and position
-  // of the menu, and by us to give back the focus to the trigger when the menu
-  // is closed (https://jira.mongodb.org/browse/PD-1674).
-  const menuTriggerRef = useRef<HTMLButtonElement | null>(null);
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-  const onClick = useCallback(
-    (evt) => {
-      evt.stopPropagation();
-      if (evt.currentTarget.dataset.menuitem) {
-        setIsMenuOpen(false);
-        // Workaround for https://jira.mongodb.org/browse/PD-1674
-        menuTriggerRef.current?.focus();
-      }
-      onAction(evt.currentTarget.dataset.action);
-    },
-    [onAction]
-  );
-
-  const shouldRender = isMenuOpen || (isVisible && actions.length > 0);
-
-  if (!shouldRender) {
-    return null;
-  }
-
-  return (
-    <Menu
-      open={isMenuOpen}
-      setOpen={setIsMenuOpen}
-      justify="start"
-      refEl={menuTriggerRef}
-      usePortal={usePortal}
-      data-testid={dataTestId}
-      trigger={({
-        onClick,
-        children,
-      }: {
-        onClick: React.MouseEventHandler<HTMLButtonElement>;
-        children: React.ReactNode;
-      }) => {
-        return (
-          <Button
-            {...buttonProps}
-            ref={menuTriggerRef}
-            data-testid={dataTestId ? `${dataTestId}-show-actions` : undefined}
-            onClick={(evt) => {
-              evt.stopPropagation();
-              onClick && onClick(evt);
-            }}
-            rightGlyph={<Icon glyph={'CaretDown'} />}
-            title={buttonText}
-          >
-            <span className={hideOnNarrow ? hiddenOnNarrowStyles : undefined}>
-              {buttonText}
-            </span>
-            {children}
-          </Button>
-        );
-      }}
-    >
-      {actions.map((menuAction, idx) => {
-        if (isSeparatorMenuAction(menuAction)) {
-          return <MenuSeparator key={`separator-${idx}`} />;
-        }
-
-        const { action, label, icon } = menuAction;
-        return (
-          <MenuItem
-            active={activeAction === action}
-            key={action}
-            data-testid={actionTestId<Action>(dataTestId, action)}
-            data-action={action}
-            data-menuitem={true}
-            glyph={<ActionGlyph glyph={icon} size={iconSize} />}
-            onClick={onClick}
-          >
-            {label}
-          </MenuItem>
-        );
-      })}
-    </Menu>
-  );
 }
