+++ packages/compass-components/src/components/actions/utils.ts
@@ -1,0 +1,6 @@
+export function actionTestId<Action extends string>(
+  dataTestId: string | undefined,
+  action: Action
+) {
+  return dataTestId ? `${dataTestId}-${action}-action` : undefined;
+}
